+++ examples/bitcoin/CHANGELOG.md
@@ -1,5 +1,32 @@
 # xchainjs-bitcoin
 
+## 1.0.8
+
+### Patch Changes
+
+- @xchainjs/xchain-bitcoin@2.0.8
+
+## 1.0.7
+
+### Patch Changes
+
+- @xchainjs/xchain-bitcoin@2.0.7
+
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [59a4a07]
+  - @xchainjs/xchain-bitcoin@2.0.6
+  - @xchainjs/xchain-util@2.0.4
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [ba9247b]
+  - @xchainjs/xchain-bitcoin@2.0.5
+
 ## 1.0.4
 
 ### Patch Changes