+++ examples/prepareTx/CHANGELOG.md
@@ -1,5 +1,40 @@
 # xchainjs-prepare-tx
 
+## 2.0.12
+
+### Patch Changes
+
+- @xchainjs/xchain-bitcoin@2.0.8
+- @xchainjs/xchain-client@2.0.8
+- @xchainjs/xchain-ethereum@2.0.12
+
+## 2.0.11
+
+### Patch Changes
+
+- @xchainjs/xchain-bitcoin@2.0.7
+- @xchainjs/xchain-client@2.0.7
+- @xchainjs/xchain-ethereum@2.0.11
+
+## 2.0.10
+
+### Patch Changes
+
+- Updated dependencies [59a4a07]
+  - @xchainjs/xchain-bitcoin@2.0.6
+  - @xchainjs/xchain-client@2.0.6
+  - @xchainjs/xchain-ethereum@2.0.10
+  - @xchainjs/xchain-util@2.0.4
+
+## 2.0.9
+
+### Patch Changes
+
+- Updated dependencies [ba9247b]
+  - @xchainjs/xchain-bitcoin@2.0.5
+  - @xchainjs/xchain-client@2.0.5
+  - @xchainjs/xchain-ethereum@2.0.9
+
 ## 2.0.8
 
 ### Patch Changes