+++ packages/xchain-mayamidgard-query/CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+## 1.0.9
+
+### Patch Changes
+
+- @xchainjs/xchain-client@2.0.8
+
+## 1.0.8
+
+### Patch Changes
+
+- @xchainjs/xchain-client@2.0.7
+
+## 1.0.7
+
+### Patch Changes
+
+- 59a4a07: Fix vulnerability form-data
+- Updated dependencies [59a4a07]
+  - @xchainjs/xchain-client@2.0.6
+  - @xchainjs/xchain-mayamidgard@1.0.4
+  - @xchainjs/xchain-util@2.0.4
+
+## 1.0.6
+
+### Patch Changes
+
+- ba9247b: use a semver version range for dependencies
+- Updated dependencies [ba9247b]
+  - @xchainjs/xchain-mayamidgard@1.0.3
+  - @xchainjs/xchain-client@2.0.5
+
 ## 1.0.5
 
 ### Patch Changes