+++ packages/xchain-utxo/CHANGELOG.md
@@ -1,5 +1,37 @@
 # Changelog
 
+## 2.0.8
+
+### Patch Changes
+
+- @xchainjs/xchain-client@2.0.8
+- @xchainjs/xchain-utxo-providers@2.0.8
+
+## 2.0.7
+
+### Patch Changes
+
+- @xchainjs/xchain-client@2.0.7
+- @xchainjs/xchain-utxo-providers@2.0.7
+
+## 2.0.6
+
+### Patch Changes
+
+- 59a4a07: Fix vulnerability form-data
+- Updated dependencies [59a4a07]
+  - @xchainjs/xchain-client@2.0.6
+  - @xchainjs/xchain-util@2.0.4
+  - @xchainjs/xchain-utxo-providers@2.0.6
+
+## 2.0.5
+
+### Patch Changes
+
+- Updated dependencies [ba9247b]
+  - @xchainjs/xchain-utxo-providers@2.0.5
+  - @xchainjs/xchain-client@2.0.5
+
 ## 2.0.4
 
 ### Patch Changes